+++ crates/sncast/src/response/show_config.rs
@@ -1,3 +1,4 @@
+use crate::Network;
 use crate::helpers::block_explorer;
 use crate::response::cast_message::SncastCommandMessage;
 use camino::Utf8PathBuf;
@@ -10,6 +11,7 @@ pub struct ShowConfigResponse {
     pub profile: Option<String>,
     pub chain_id: Option<String>,
     pub rpc_url: Option<Url>,
+    pub network: Option<Network>,
     pub account: Option<String>,
     pub accounts_file_path: Option<Utf8PathBuf>,
     pub keystore: Option<Utf8PathBuf>,
@@ -30,6 +32,9 @@ impl SncastCommandMessage for ShowConfigResponse {
             })
             .if_some(self.rpc_url.as_ref(), |b, rpc_url| {
                 b.field("RPC URL", rpc_url.as_ref())
+            })
+            .if_some(self.network.as_ref(), |b, network| {
+                b.field("Network", network.to_string().as_ref())
             })
             .if_some(self.account.as_ref(), |b, account| {
                 b.field("Account", account)